@@ -1,6 +1,6 @@
 module proto.zip/studio/validate
 
-go 1.20
+go 1.22
 
 require (
 	golang.org/x/net v0.15.0
