@@ -1,6 +1,7 @@
 package net_test
 
 import (
+	"context"
 	"strings"
 	"testing"
 
@@ -13,65 +14,74 @@ import (
 // - Default configuration doesn't return errors on valid value.
 // - Implements interface.
 func TestDomainRuleSet(t *testing.T) {
-	d, err := net.NewDomain().Validate("example.com")
+	// Prepare the output variable for Apply
+	var output string
+
+	example := "example.com"
+
+	// Apply with a valid domain string
+	err := net.Domain().Apply(context.TODO(), example, &output)
 
 	if err != nil {
 		t.Errorf("Expected errors to be empty, got: %s", err)
 		return
 	}
 
-	if d != "example.com" {
+	if output != example {
 		t.Error("Expected test domain to be returned")
 		return
 	}
 
-	ok := testhelpers.CheckRuleSetInterface[string](net.NewDomain())
+	// Check if the rule set implements the expected interface
+	ok := testhelpers.CheckRuleSetInterface[string](net.Domain())
 	if !ok {
 		t.Error("Expected rule set to be implemented")
 		return
 	}
+
+	testhelpers.MustApplyTypes[string](t, net.Domain(), example)
 }
 
 // Requirements:
 // - Segments (labels) cannot exceed 63 characters
 // See: RFC 1035
 func TestDomainSegmentLength(t *testing.T) {
-	ruleSet := net.NewDomain().Any()
+	ruleSet := net.Domain().Any()
 
 	okLabel := strings.Repeat("a", 63)
 	badLabel := strings.Repeat("a", 64)
 
-	testhelpers.MustBeValid(t, ruleSet, okLabel+".com", okLabel+".com")
-	testhelpers.MustBeInvalid(t, ruleSet, badLabel+".com", errors.CodePattern)
+	testhelpers.MustApply(t, ruleSet, okLabel+".com")
+	testhelpers.MustNotApply(t, ruleSet, badLabel+".com", errors.CodePattern)
 }
 
 // Requirements:
 // - Errors when string cannot be encoded as punycode
 func TestDomainPunycodeError(t *testing.T) {
-	ruleSet := net.NewDomain().Any()
+	ruleSet := net.Domain().Any()
 
 	// idna: invalid label "é"
 	str := "example.xn--é.com"
-	testhelpers.MustBeInvalid(t, ruleSet, str+".com", errors.CodePattern)
+	testhelpers.MustNotApply(t, ruleSet, str+".com", errors.CodePattern)
 }
 
 // Requirements:
 // - Errors when domain is too long
 // - errors.CodeMax is returned
 func TestDomainLength(t *testing.T) {
-	ruleSet := net.NewDomain().Any()
+	ruleSet := net.Domain().Any()
 
 	str := strings.Repeat(strings.Repeat("a", 32), 9)
-	testhelpers.MustBeInvalid(t, ruleSet, str+".com", errors.CodeMax)
+	testhelpers.MustNotApply(t, ruleSet, str+".com", errors.CodeMax)
 }
 
 // Requirements:
 // - Errors when input is not a string
 // - errors.CodeType is returned
 func TestDomainType(t *testing.T) {
-	ruleSet := net.NewDomain().Any()
+	ruleSet := net.Domain().Any()
 
-	testhelpers.MustBeInvalid(t, ruleSet, 123, errors.CodeType)
+	testhelpers.MustNotApply(t, ruleSet, 123, errors.CodeType)
 }
 
 // Requirements:
@@ -79,7 +89,7 @@ func TestDomainType(t *testing.T) {
 // - Required flag can be read.
 // - Required flag defaults to false.
 func TestDomainRequired(t *testing.T) {
-	ruleSet := net.NewDomain()
+	ruleSet := net.Domain()
 
 	if ruleSet.Required() {
 		t.Error("Expected rule set to not be required")
@@ -93,28 +103,40 @@ func TestDomainRequired(t *testing.T) {
 }
 
 func TestDomainCustom(t *testing.T) {
-	_, err := net.NewDomain().
-		WithRuleFunc(testhelpers.MockCustomRule("example.com", 1)).
-		Validate("example.com")
+	mock := testhelpers.NewMockRuleWithErrors[string](1)
+
+	// Prepare the output variable for Apply
+	var output string
+
+	// Apply with a mock rule that should trigger an error
+	err := net.Domain().
+		WithRuleFunc(mock.Function()).
+		Apply(context.TODO(), "example.com", &output)
 
 	if err == nil {
 		t.Error("Expected errors to not be empty")
 		return
 	}
 
-	expected := "example.com"
+	if mock.EvaluateCallCount() != 1 {
+		t.Errorf("Expected rule to be called 1 time, got %d", mock.EvaluateCallCount())
+		return
+	}
 
-	actual, err := net.NewDomain().
-		WithRuleFunc(testhelpers.MockCustomRule(expected, 0)).
-		Validate("example.com")
+	rule := testhelpers.NewMockRule[string]()
+
+	// Apply with a mock rule that should pass without errors
+	err = net.Domain().
+		WithRuleFunc(rule.Function()).
+		Apply(context.TODO(), "example.com", &output)
 
 	if err != nil {
 		t.Error("Expected errors to be empty")
 		return
 	}
 
-	if expected != actual {
-		t.Errorf("Expected '%s' to equal '%s'", actual, expected)
+	if c := rule.EvaluateCallCount(); c != 1 {
+		t.Errorf("Expected rule to be called once, got %d", c)
 		return
 	}
 }
@@ -122,7 +144,7 @@ func TestDomainCustom(t *testing.T) {
 // Requirements:
 // - Serializes to WithRequired()
 func TestDomainRequiredString(t *testing.T) {
-	ruleSet := net.NewDomain().WithRequired()
+	ruleSet := net.Domain().WithRequired()
 
 	expected := "DomainRuleSet.WithRequired()"
 	if s := ruleSet.String(); s != expected {
