@@ -1,5 +1,5 @@
 // Code generated by running "go generate". DO NOT EDIT.
-// Version 2023101002, Last Updated Wed Oct 11 07:07:02 2023 UTC
+// Version 2024071200, Last Updated Fri Jul 12 07:07:01 2024 UTC
 
 package net
 
@@ -91,7 +91,6 @@ var TLDs = []string{
 	"AUTHOR",
 	"AUTO",
 	"AUTOS",
-	"AVIANCA",
 	"AW",
 	"AWS",
 	"AX",
@@ -102,7 +101,6 @@ var TLDs = []string{
 	"BABY",
 	"BAIDU",
 	"BANAMEX",
-	"BANANAREPUBLIC",
 	"BAND",
 	"BANK",
 	"BAR",
@@ -219,7 +217,6 @@ var TLDs = []string{
 	"CBA",
 	"CBN",
 	"CBRE",
-	"CBS",
 	"CC",
 	"CD",
 	"CENTER",
@@ -248,7 +245,6 @@ var TLDs = []string{
 	"CITI",
 	"CITIC",
 	"CITY",
-	"CITYEATS",
 	"CK",
 	"CL",
 	"CLAIMS",
@@ -269,7 +265,6 @@ var TLDs = []string{
 	"COLLEGE",
 	"COLOGNE",
 	"COM",
-	"COMCAST",
 	"COMMBANK",
 	"COMMUNITY",
 	"COMPANY",
@@ -387,7 +382,6 @@ var TLDs = []string{
 	"ESQ",
 	"ESTATE",
 	"ET",
-	"ETISALAT",
 	"EU",
 	"EUROVISION",
 	"EUS",
@@ -450,7 +444,6 @@ var TLDs = []string{
 	"FRESENIUS",
 	"FRL",
 	"FROGANS",
-	"FRONTDOOR",
 	"FRONTIER",
 	"FTR",
 	"FUJITSU",
@@ -522,7 +515,6 @@ var TLDs = []string{
 	"GS",
 	"GT",
 	"GU",
-	"GUARDIAN",
 	"GUCCI",
 	"GUGE",
 	"GUIDE",
@@ -651,7 +643,6 @@ var TLDs = []string{
 	"KIA",
 	"KIDS",
 	"KIM",
-	"KINDER",
 	"KINDLE",
 	"KITCHEN",
 	"KIWI",
@@ -817,7 +808,6 @@ var TLDs = []string{
 	"NAB",
 	"NAGOYA",
 	"NAME",
-	"NATURA",
 	"NAVY",
 	"NBA",
 	"NC",
@@ -866,7 +856,6 @@ var TLDs = []string{
 	"OKINAWA",
 	"OLAYAN",
 	"OLAYANGROUP",
-	"OLDNAVY",
 	"OLLO",
 	"OM",
 	"OMEGA",
@@ -993,7 +982,6 @@ var TLDs = []string{
 	"RIO",
 	"RIP",
 	"RO",
-	"ROCHER",
 	"ROCKS",
 	"RODEO",
 	"ROGERS",
@@ -1028,7 +1016,6 @@ var TLDs = []string{
 	"SBI",
 	"SBS",
 	"SC",
-	"SCA",
 	"SCB",
 	"SCHAEFFLER",
 	"SCHMIDT",
@@ -1057,7 +1044,6 @@ var TLDs = []string{
 	"SH",
 	"SHANGRILA",
 	"SHARP",
-	"SHAW",
 	"SHELL",
 	"SHIA",
 	"SHIKSHA",
@@ -1066,7 +1052,6 @@ var TLDs = []string{
 	"SHOPPING",
 	"SHOUJI",
 	"SHOW",
-	"SHOWTIME",
 	"SI",
 	"SILK",
 	"SINA",
@@ -1247,7 +1232,6 @@ var TLDs = []string{
 	"VLAANDEREN",
 	"VN",
 	"VODKA",
-	"VOLKSWAGEN",
 	"VOLVO",
 	"VOTE",
 	"VOTING",
@@ -1291,7 +1275,6 @@ var TLDs = []string{
 	"WTF",
 	"XBOX",
 	"XEROX",
-	"XFINITY",
 	"XIHUAN",
 	"XIN",
 	"XN--11B4C3D",
@@ -1382,7 +1365,6 @@ var TLDs = []string{
 	"XN--MGBA3A3EJT",
 	"XN--MGBA3A4F16A",
 	"XN--MGBA7C0BBN0A",
-	"XN--MGBAAKC7DVF",
 	"XN--MGBAAM7A8H",
 	"XN--MGBAB2BD",
 	"XN--MGBAH1A3HJKRD",
