@@ -15,14 +15,14 @@ type minTimeRule struct {
 }
 
 // Evaluate takes a context and integer value and returns an error if it is not equal or later than than the specified value.
-func (rule *minTimeRule) Evaluate(ctx context.Context, value time.Time) (time.Time, errors.ValidationErrorCollection) {
+func (rule *minTimeRule) Evaluate(ctx context.Context, value time.Time) errors.ValidationErrorCollection {
 	if value.Before(rule.min) {
-		return value, errors.Collection(
+		return errors.Collection(
 			errors.Errorf(errors.CodeMin, ctx, "field must be on or after %s", rule.min),
 		)
 	}
 
-	return value, nil
+	return nil
 }
 
 // Conflict returns true for any minimum rule.
@@ -39,13 +39,6 @@ func (rule *minTimeRule) String() string {
 
 // WithMin returns a new child RuleSet that is constrained to the provided minimum time value.
 func (v *TimeRuleSet) WithMin(min time.Time) *TimeRuleSet {
-	return v.WithRule(&minTimeRule{
-		min,
-	})
-}
-
-// WithMin returns a new child RuleSet that is constrained to the provided minimum time value.
-func (v *TimeStringRuleSet) WithMin(min time.Time) *TimeStringRuleSet {
 	return v.WithRule(&minTimeRule{
 		min,
 	})
