@@ -10,6 +10,16 @@ import (
 	"proto.zip/studio/validate/pkg/rules"
 )
 
+// neverAssignable if an interface with a private method making it impossible to assign to while.
+// still satisfying interface{} type checks.
+// Used by MustApplyTypes and not exported.
+type neverAssignable interface{ priv() }
+
+// neverAssignableI is an implementation of neverAssignable for use in MustApplyTypes
+type neverAssignableImpl struct{ privProp int }
+
+func (na *neverAssignableImpl) priv() {}
+
 // CheckRuleSetInterface checks to see if the RuleSet interface is implemented for an interface and returns true if it is.
 func CheckRuleSetInterface[T any](v any) bool {
 	_, ok := v.(rules.RuleSet[T])
@@ -22,80 +32,195 @@ func CheckRuleInterface[T any](v any) bool {
 	return ok
 }
 
-// MockCustomRule is a mock implementation of the Rule interface that can be used for testing.
-// It accepts a return value to return from the rule and also a number of errors to return.
-//
-// If errorCount is 0 than nil is returned for errors.
-// The return value of this function is a Rule function.
-func MockCustomRule[T any](retval T, errorCount int) func(_ context.Context, _ T) (T, errors.ValidationErrorCollection) {
-	var errs errors.ValidationErrorCollection
-
-	if errorCount > 0 {
-		errs = make(errors.ValidationErrorCollection, errorCount)
-
-		for i := 0; i < errorCount; i++ {
-			errs[i] = errors.Errorf(errors.CodeUnknown, context.Background(), "test")
-		}
-	}
-
-	return func(_ context.Context, _ T) (T, errors.ValidationErrorCollection) {
-		return retval, errs
-	}
-}
-
 // checkEqual is a simple validity function that returns true if both values are equal.
 func checkEqual(a, b any) error {
 	if a != b {
-		return fmt.Errorf("Expected output to be %v, got: %v", a, b)
+		return fmt.Errorf("expected output to be %v, got: %v", a, b)
 	}
 	return nil
 }
 
-// MustBeValidFunc is a test helper that expects a RuleSet to a nil error.
+// checkAlways is a check function that always returns nil.
+func checkAlways(_, _ any) error {
+	return nil
+}
+
+// MustApplyFunc is a test helper that expects a RuleSet to return a nil error.
 // If the error is non-nil or the check function returns an error, this function prints the error and returns it.
-func MustBeValidFunc(t testing.TB, ruleSet rules.RuleSet[any], input, expectedOutput any, fn func(a, b any) error) error {
+func MustApplyFunc(t testing.TB, ruleSet rules.RuleSet[any], input, expectedOutput any, fn func(a, b any) error) (any, error) {
 	t.Helper()
 
-	actualOutput, err := ruleSet.Validate(input)
+	// Initialize the actual output variable
+	var actualOutput any
+	err := ruleSet.Apply(context.TODO(), input, &actualOutput)
 
 	if err != nil {
 		str := "Expected error to be nil"
 
 		for _, inner := range err {
-			str = "\n  " + fmt.Sprintf("%s at %s", inner, inner.Path())
+			str += fmt.Sprintf("\n  %s at %s", inner, inner.Path())
 		}
 
 		t.Errorf(str)
-		return err
+		return actualOutput, err
 	} else if err := fn(expectedOutput, actualOutput); err != nil {
 		t.Error(err)
+		return actualOutput, err
+	}
+
+	return actualOutput, nil
+}
+
+// MustApply is a test helper that expects a RuleSet to return the input value and nil error.
+// If the error is non-nil or the expected output does not match, this function prints the error and returns it.
+func MustApply(t testing.TB, ruleSet rules.RuleSet[any], input any) (any, error) {
+	t.Helper()
+	return MustApplyFunc(t, ruleSet, input, input, checkEqual)
+}
+
+// MustApplyAny is a test helper that expects a RuleSet to finish without an error.
+func MustApplyAny(t testing.TB, ruleSet rules.RuleSet[any], input any) (any, error) {
+	t.Helper()
+	return MustApplyFunc(t, ruleSet, input, input, checkAlways)
+}
+
+// MustApplyMutation is a test helper that expects a RuleSet to return a specific value and nil error.
+// If the error is non-nil or the expected output does not match, this function prints the error and returns it.
+func MustApplyMutation(t testing.TB, ruleSet rules.RuleSet[any], input, output any) (any, error) {
+	t.Helper()
+	return MustApplyFunc(t, ruleSet, input, output, checkEqual)
+}
+
+// MustNotApply is a test helper that expects a RuleSet to return an error and checks for a specific error code.
+// If the error is nil or the code does not match, a testing error is printed and the function returns false.
+//
+// This function returns the error on "success" so that you can perform additional comparisons.
+func MustNotApply(t testing.TB, ruleSet rules.RuleSet[any], input any, errorCode errors.ErrorCode) error {
+	t.Helper()
+
+	var output any
+	err := ruleSet.Apply(context.TODO(), input, &output)
+
+	if err == nil {
+		t.Error("Expected error to not be nil")
+		return nil
+	} else if err.First().Code() != errorCode {
+		t.Errorf("Expected error code of %s, got %s (%s)", errorCode, err.First().Code(), err)
+		return nil
+	}
+
+	return err
+}
+
+// MustApplyTypes checks to make sure apply supports the various output types expected all rule sets.
+// It is recommended all RuleSet implementations pass this assertion.
+//
+// Output types tested are:
+// - Pointer to any.
+// - Pointer to correct type.
+// - Non-pointer (should error).
+// - Pointer to nil (should error).
+// - Nil (should error).
+//
+// Be sure to use an input that should not error if the types are correct.
+// Note that Apply may implement output types other than these but these are bare minimum for any public RuleSet.
+func MustApplyTypes[T any](t testing.TB, ruleSet rules.RuleSet[T], input T) {
+	t.Helper()
+
+	// Do not use MustApply and MustNotApply as these require .Any() which may invalidate the test.
+
+	// Pointer to any
+	var outputAny any
+	err := ruleSet.Apply(context.TODO(), input, &outputAny)
+	if err != nil {
+		t.Errorf("Expected error to be nil on `any` output, got: %s", err)
+	}
+
+	// Pointer to correct type
+	var outputPtr *T = new(T)
+	err = ruleSet.Apply(context.TODO(), input, outputPtr)
+	if err != nil {
+		t.Errorf("Expected error to be nil on `%T` output, got: %s", outputPtr, err)
+	}
+
+	// Non-pointer to correct type
+	var outputNonPointer T
+	err = ruleSet.Apply(context.TODO(), input, outputNonPointer)
+	if err == nil {
+		t.Errorf("Expected error to not be nil on `%T` output", outputNonPointer)
+	} else if code := err.First().Code(); code != errors.CodeInternal {
+		t.Errorf("Expected error code to be %s (errors.CodeInternal) on `%T` output, got: %s", errors.CodeInternal, outputNonPointer, code)
+	}
+
+	// Pointer to nil
+	var outputPointerToNil *T
+	err = ruleSet.Apply(context.TODO(), input, outputPointerToNil)
+	if err == nil {
+		t.Error("Expected error to not be nil on pointer to `nil` output")
+	} else if code := err.First().Code(); code != errors.CodeInternal {
+		t.Errorf("Expected error code to be %s (errors.CodeInternal) on pointer to `nil` output, got: %s", errors.CodeInternal, code)
+	}
+
+	// Incompatible type
+	// We must assign a &neverAssignableImpl{} to avoid false errors because the pointer was nil
+	var outputIncompatible neverAssignable = &neverAssignableImpl{privProp: 1}
+	outputIncompatible.priv()
+	err = ruleSet.Apply(context.TODO(), input, outputIncompatible)
+	if err == nil {
+		t.Error("Expected error to not be nil on incompatible output")
+	} else if code := err.First().Code(); code != errors.CodeInternal {
+		t.Errorf("Expected error code to be %s (errors.CodeInternal) on incompatible output, got: %s", errors.CodeInternal, code)
+	}
+
+	// Nil value
+	err = ruleSet.Apply(context.TODO(), input, nil)
+	if err == nil {
+		t.Error("Expected error to not be nil on `nil` output")
+	} else if code := err.First().Code(); code != errors.CodeInternal {
+		t.Errorf("Expected error code to be %s (errors.CodeInternal) on `nil` output, got: %s", errors.CodeInternal, code)
+	}
+}
+
+// MustEvaluate is a test helper that expects a Rule to return nil for an error.
+// If the error is non-nil, this function prints the error and returns it.
+//
+// WrapAnyRuleSet calls Apply because of the less strict input type, so when testing Evaluate it is important
+// to not use .Any() with this test.
+func MustEvaluate[T any](t testing.TB, rule rules.Rule[T], input T) error {
+	t.Helper()
+
+	err := rule.Evaluate(context.TODO(), input)
+
+	if err != nil {
+		str := "Expected error to be nil"
+
+		for _, inner := range err {
+			str += fmt.Sprintf("\n  %s at %s", inner, inner.Path())
+		}
+
+		t.Errorf(str)
 		return err
 	}
 
 	return nil
 }
 
-// MustBeValid is a test helper that expects a RuleSet to return a specific value and nil error.
-// If the error is non-nil or the expected output does not match, this function prints the error and returns it.
-func MustBeValid(t testing.TB, ruleSet rules.RuleSet[any], input, expectedOutput any) error {
-	t.Helper()
-	return MustBeValidFunc(t, ruleSet, input, expectedOutput, checkEqual)
-}
-
-// MustBeInvalid is a test helper that expects a RuleSet to return an error and checks for a specific error code.
+// MustNotEvaluate is a test helper that expects a Rule to return an error and checks for a specific error code.
 // If the error is nil or the code does not match, a testing error is printed and the function returns false.
 //
 // This function returns the error on "success" so that you can perform additional comparisons.
-func MustBeInvalid(t testing.TB, ruleSet rules.RuleSet[any], input any, errorCode errors.ErrorCode) error {
+//
+// WrapAnyRuleSet calls Apply because of the less strict input type, so when testing Evaluate it is important
+// to not use .Any() with this test.
+func MustNotEvaluate[T any](t testing.TB, rule rules.Rule[T], input T, errorCode errors.ErrorCode) error {
 	t.Helper()
-
-	_, err := ruleSet.Validate(input)
+	err := rule.Evaluate(context.TODO(), input)
 
 	if err == nil {
 		t.Error("Expected error to not be nil")
 		return nil
 	} else if err.First().Code() != errorCode {
-		t.Errorf("Expected error code of %s got %s (%s)", errorCode, err.First().Code(), err)
+		t.Errorf("Expected error code of %s, got %s (%s)", errorCode, err.First().Code(), err)
 		return nil
 	}
 
